+++ src/Blamantic/Abstractions/BlamenticComponentBase.cs
@@ -1,4 +1,7 @@
-﻿using YoiBlazor;
+﻿using System.Reflection;
+using Microsoft.AspNetCore.Components;
+using Microsoft.AspNetCore.Components.Rendering;
+using YoiBlazor;
 
 namespace BlamanticUI.Abstractions
 {
@@ -13,6 +16,19 @@ namespace BlamanticUI.Abstractions
         /// </summary>
         protected BlamanticComponentBase() : base()
         {
+        }
+
+        /// <summary>
+        /// Throws the <see cref="CascadingComponentException"/> when specify parent component is null.
+        /// </summary>
+        /// <param name="parent">The parent component witch is cascaded from parent.</param>
+        /// <exception cref="CascadingComponentException">Current component must inside <paramref name="parent"/> component.</exception>
+        protected virtual void ThrowParentIsNull(IComponent parent)
+        {
+            if(parent is null)
+            {
+                throw new CascadingComponentException(parent, this);
+            }
         }
 
     }