+++ src/Blamantic/Components/Table/TableRow.cs
@@ -1,10 +1,8 @@
 ﻿
-using System.Threading.Tasks;
-
 using BlamanticUI.Abstractions;
 
 using Microsoft.AspNetCore.Components;
-
+using Microsoft.AspNetCore.Components.Rendering;
 using YoiBlazor;
 
 namespace BlamanticUI
@@ -16,13 +14,17 @@ namespace BlamanticUI
     /// <seealso cref="BlamanticUI.Abstractions.IHasState" />
     /// <seealso cref="BlamanticUI.Abstractions.IHasActive" />
     /// <seealso cref="BlamanticUI.Abstractions.IHasDisabled" />
-    [HtmlTag("tr")]
-    public class Tr : BlamanticChildContentComponentBase,IHasState,IHasActive,IHasDisabled,IHasColor,IHasHorizontalAlignment,IHasVerticalAlignment
+    public class TableRow : BlamanticChildContentComponentBase,IHasState,IHasActive,IHasDisabled,IHasColor,IHasHorizontalAlignment,IHasVerticalAlignment
     {
+        /// <summary>
+        /// Gets or sets the cascading table.
+        /// </summary>
+        [CascadingParameter] Table CascadingTable { get; set; }
+
         /// <summary>
         /// Gets or sets the marker style of border in cell.
         /// </summary>
-        [Parameter] [CssClass(" marked",Suffix =true)] public HorizontalPosition? Marked { get; set; }
+        [Parameter] [CssClass(" marked", Suffix = true)] public HorizontalPosition? Marked { get; set; }
         /// <summary>
         /// Gets or sets the state of row.
         /// </summary>
@@ -61,5 +63,27 @@ namespace BlamanticUI
         /// Gets or sets the horizontal alignment of text.
         /// </summary>
         [Parameter]public HorizontalAlignment? HorizontalAlignment { get; set; }
+
+        /// <summary>
+        /// Method invoked when the component is ready to start, having received its
+        /// initial parameters from its parent in the render tree.
+        /// </summary>
+        protected override void OnInitialized()
+        {
+            ThrowParentIsNull(CascadingTable);
+            base.OnInitialized();
+        }
+
+        /// <summary>
+        /// Renders the component to the supplied <see cref="T:Microsoft.AspNetCore.Components.Rendering.RenderTreeBuilder" />.
+        /// </summary>
+        /// <param name="builder">A <see cref="T:Microsoft.AspNetCore.Components.Rendering.RenderTreeBuilder" /> that will receive the render output.</param>
+        protected override void BuildRenderTree(RenderTreeBuilder builder)
+        {
+            builder.OpenElement(0, "tr");
+            AddCommonAttributes(builder);
+            builder.BuildCascadingValueComponent<TableRow>(this, ChildContent);
+            builder.CloseElement();
+        }
     }
 }